Add pytest smoke target
