+++ pyproject.toml
@@ -46,6 +46,7 @@ markers = [
     "limitador: Test is using Limitador features",
     "tlspolicy: Test is using TLSPolicy",
     "dnspolicy: Test is using DNSPolicy",
+    "smoke: Build verification test",
     "disruptive: Test is disruptive",
     "multicluster: Test is specifc to Multicluster deployment",
 ]
