+++ testsuite/tests/singlecluster/limitador/test_basic_limit.py
@@ -12,7 +12,7 @@ pytestmark = [pytest.mark.kuadrant_only, pytest.mark.limitador]
 @pytest.fixture(
     scope="module",
     params=[
-        pytest.param(Limit(2, "15s"), id="2 requests every 15 sec"),
+        pytest.param(Limit(2, "15s"), id="2 requests every 15 sec", marks=[pytest.mark.smoke]),
         pytest.param(Limit(5, "10s"), id="5 requests every 10 sec"),
         pytest.param(Limit(3, "5s"), id="3 request every 5 sec"),
     ],
